requirements.txt
ultralytics
streamlit
pillow
numpy
opencv-python